refactor: Modernise Status
The foundations of the status page were written 10 years ago, a number
of the components are no longer supported, and there are some rough
edges.

This serves as a rewrite in a current react version, combined with
tailwind css and switching bootstrap for shad-cn.

It also addresses #1 by adjusting the layout when viewed on a mobile
device.